Move test to test_dataloaders and add extra test
